requirements.txt
absl-py==0.12.0
appdirs==1.4.3
astunparse==1.6.3
cachetools==4.2.2
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
dbus-python==1.2.16
distlib==0.3.0
distro==1.4.0
easydict==1.9
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.5.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
linecache2==1.0.0
Markdown==3.3.4
matplotlib==3.4.2
more-itertools==4.2.0
numpy==1.19.5
oauthlib==3.1.0
onnx-graphsurgeon==0.2.6
opt-einsum==3.3.0
pbr==5.4.5
Pillow==8.3.1
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyGObject==3.36.0
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
Shapely==1.7.1
six==1.15.0
ssh-import-id==5.10
stevedore==1.32.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tensorrt==8.0.0.3
termcolor==1.1.0
testresources==2.0.0
tqdm==4.62.0
traceback2==1.4.0
typing-extensions==3.7.4.3
unittest2==1.1.0
urllib3==1.25.8
virtualenv==20.0.17
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.0.0
